Processing Web Content On WID Based ADFS Servers
When using ADFS 2012 R2 with WID, a bug exist where specific web content does not replicate to secondary ADFS servers. This script helps in configuring the custom web content on secondary ADFS servers!

Make sure to read the blog post first before using this! Test in testlab first!!!

For a detailed explanation and usage of this script, please go to:

(2016-07-24) Fixing Web Content Data In ADFS 2012 R2 (v3.0) When Leveraging WID As A Database Store